struct: add modules for trees and nodes refactor: update Copyright header
