+++ app/src/main/kotlin/trees/KeyValue.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) 2023 teemEight
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package trees
 
 class KeyValue<K : Comparable<K>, V>(private val key: K, private var value: V?) : Comparable<KeyValue<K, V>> {
