+++ app/src/main/kotlin/trees/interfaces/Node.kt
@@ -1,7 +1,9 @@
 /*
- * Copyright 2023 teemEight
+ * Copyright (c) 2023 teemEight
  * SPDX-License-Identifier: Apache-2.0
  */
+
+package trees.interfaces
 
 interface Node<T : Comparable<T>, Subtype : Node<T, Subtype>> : Comparable<Node<T, Subtype>> {
     var data: T
