+++ app/src/main/kotlin/trees/interfaces/Tree.kt
@@ -1,4 +1,9 @@
-package trees
+/*
+ * Copyright (c) 2023 teemEight
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package trees.interfaces
 
 interface Tree<T : Comparable<T>> {
     fun add(data: T)
