+++ app/src/main/kotlin/trees/nodes/AVLNode.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2023 teemEight
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package trees.nodes
+
+class AVLNode {
+}
