+++ app/src/main/kotlin/trees/trees/AVLTree.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2023 teemEight
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package trees.trees
+
+class AVLTree {
+}
